swap to stardust-2 base

+++ MMOCoreORB/src/server/zone/objects/player/sessions/InterplanetarySurveyDroidSessionImplementation.cpp
@@ -150,15 +150,15 @@ void InterplanetarySurveyDroidSessionImplementation::handleMenuSelect(CreatureOb
 		}
 
 		float quality = component->getAttributeValue("mechanism_quality");
-		unsigned long chosen = droidSuiBox->getMenuObjectID(menuID);
+		uint64 chosen = droidSuiBox->getMenuObjectID(menuID);
 		this->targetPlanet = pl->getZoneServer()->getResourceManager()->getPlanetByIndex(chosen);
-		int duration = 1000 * (3600 - (27 * quality));
-		int minutes = duration/60000;
+		int duration = 30000;
+		int minutes = duration/1000;
 
 		StringBuffer buffer;
 		buffer << "Droid sent, ETA for the report is ";
 		buffer << minutes;
-		buffer << " minutes.";
+		buffer << " seconds.";
 		pl->sendSystemMessage(buffer.toString());
 
 		// Create a bogus task to run to show the output to the console
@@ -179,8 +179,13 @@ void InterplanetarySurveyDroidSessionImplementation::handleMenuSelect(CreatureOb
 
 		ObjectManager::instance()->persistObject(data, 1, "surveys");
 
-		tool->destroyObjectFromWorld(true);
-		tool->destroyObjectFromDatabase(true);
+		// Rather than always using a tool per survey droid, simply damage it.
+						tool->setConditionDamage(tool->getConditionDamage() + 40, true);
+						// If the damage is greater than the remaining condition, trash it.
+						if (tool->isDestroyed()) {
+							tool->destroyObjectFromWorld(true);
+							tool->destroyObjectFromDatabase(true);
+						}
 
 		tangibleObject->decreaseUseCount();
 		cancelSession();
